Foram criados o cpp e h da classe Pessoa os métodos e atributos foram inicializados. os métodos estão sendo usado no main

+++ pessoa/Pessoa/Pessoa.h
@@ -9,7 +9,7 @@ public:
     void almentarFelicidade();
     
 private:
-    int altura;
+    double altura;
     int idade;
     int niveldeFelicidade;
 };
